Fix #15 "@daily @hourly etc not run"
Initialisation for day-of-week(dow) is missing for jobs that use
one of the aliases such as @hourly, @daily, @weekly.

In TestJobs() (database:1032) and ArmJob() (database.c:1105) there is
a check that involves day-of-week that consequently fails, so the job
will never be run.

Add the missing initialisation.

Credit for debugging/fixing this goes to jensrenner1980 @ github
Fixes Issue #15, introduced by fix for Issue #3